# book-server/src/main/java/org/openapitools/api/BooksApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.8.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.openapitools.api;

import org.openapitools.model.Book;
import org.openapitools.model.BookInput;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-04-18T17:08:17.792697300+08:00[Asia/Shanghai]", comments = "Generator version: 7.8.0")
@Validated
@Tag(name = "books", description = "the books API")
public interface BooksApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * DELETE /books/{bookId} : 删除图书
     * 删除指定ID的图书
     *
     * @param bookId  (required)
     * @return 图书删除成功 (status code 204)
     *         or 图书不存在 (status code 404)
     *         or 服务器内部错误 (status code 500)
     */
    @Operation(
        operationId = "booksBookIdDelete",
        summary = "删除图书",
        description = "删除指定ID的图书",
        responses = {
            @ApiResponse(responseCode = "204", description = "图书删除成功"),
            @ApiResponse(responseCode = "404", description = "图书不存在"),
            @ApiResponse(responseCode = "500", description = "服务器内部错误")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/books/{bookId}"
    )
    
    default ResponseEntity<Void> booksBookIdDelete(
        @Min(1) @Parameter(name = "bookId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("bookId") Integer bookId
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /books/{bookId} : 根据ID获取图书
     * 返回指定ID的图书详情
     *
     * @param bookId 图书ID (required)
     * @return 成功获取图书 (status code 200)
     *         or 图书不存在 (status code 404)
     *         or 服务器内部错误 (status code 500)
     */
    @Operation(
        operationId = "booksBookIdGet",
        summary = "根据ID获取图书",
        description = "返回指定ID的图书详情",
        responses = {
            @ApiResponse(responseCode = "200", description = "成功获取图书", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Book.class))
            }),
            @ApiResponse(responseCode = "404", description = "图书不存在"),
            @ApiResponse(responseCode = "500", description = "服务器内部错误")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/books/{bookId}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<Book> booksBookIdGet(
        @Min(1) @Parameter(name = "bookId", description = "图书ID", required = true, in = ParameterIn.PATH) @PathVariable("bookId") Integer bookId
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"id\" : 1, \"title\" : \"深入理解计算机系统\", \"author\" : \"Randal E. Bryant\", \"isbn\" : \"9787111321330\", \"publishedDate\" : \"2011-01-01\", \"price\" : 99.0, \"stock\" : 50 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * PUT /books/{bookId} : 更新图书信息
     * 更新指定ID的图书信息
     *
     * @param bookId  (required)
     * @param bookInput  (required)
     * @return 图书更新成功 (status code 200)
     *         or 图书不存在 (status code 404)
     *         or 请求参数无效 (status code 400)
     *         or 服务器内部错误 (status code 500)
     */
    @Operation(
        operationId = "booksBookIdPut",
        summary = "更新图书信息",
        description = "更新指定ID的图书信息",
        responses = {
            @ApiResponse(responseCode = "200", description = "图书更新成功", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Book.class))
            }),
            @ApiResponse(responseCode = "404", description = "图书不存在"),
            @ApiResponse(responseCode = "400", description = "请求参数无效"),
            @ApiResponse(responseCode = "500", description = "服务器内部错误")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/books/{bookId}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Book> booksBookIdPut(
        @Min(1) @Parameter(name = "bookId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("bookId") Integer bookId,
        @Parameter(name = "BookInput", description = "", required = true) @Valid @RequestBody BookInput bookInput
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"id\" : 1, \"title\" : \"深入理解计算机系统\", \"author\" : \"Randal E. Bryant\", \"isbn\" : \"9787111321330\", \"publishedDate\" : \"2011-01-01\", \"price\" : 99.0, \"stock\" : 50 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /books : 获取所有图书
     * 返回系统中的所有图书列表
     *
     * @return 成功获取图书列表 (status code 200)
     *         or 服务器内部错误 (status code 500)
     */
    @Operation(
        operationId = "booksGet",
        summary = "获取所有图书",
        description = "返回系统中的所有图书列表",
        responses = {
            @ApiResponse(responseCode = "200", description = "成功获取图书列表", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Book.class)))
            }),
            @ApiResponse(responseCode = "500", description = "服务器内部错误")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/books",
        produces = { "application/json" }
    )
    
    default ResponseEntity<List<Book>> booksGet(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"id\" : 1, \"title\" : \"深入理解计算机系统\", \"author\" : \"Randal E. Bryant\", \"isbn\" : \"9787111321330\", \"publishedDate\" : \"2011-01-01\", \"price\" : 99.0, \"stock\" : 50 }, { \"id\" : 1, \"title\" : \"深入理解计算机系统\", \"author\" : \"Randal E. Bryant\", \"isbn\" : \"9787111321330\", \"publishedDate\" : \"2011-01-01\", \"price\" : 99.0, \"stock\" : 50 } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /books : 添加新图书
     * 向系统中添加一本新图书
     *
     * @param bookInput  (required)
     * @return 图书创建成功 (status code 201)
     *         or 请求参数无效 (status code 400)
     *         or 服务器内部错误 (status code 500)
     */
    @Operation(
        operationId = "booksPost",
        summary = "添加新图书",
        description = "向系统中添加一本新图书",
        responses = {
            @ApiResponse(responseCode = "201", description = "图书创建成功", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Book.class))
            }),
            @ApiResponse(responseCode = "400", description = "请求参数无效"),
            @ApiResponse(responseCode = "500", description = "服务器内部错误")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/books",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Book> booksPost(
        @Parameter(name = "BookInput", description = "", required = true) @Valid @RequestBody BookInput bookInput
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"id\" : 1, \"title\" : \"深入理解计算机系统\", \"author\" : \"Randal E. Bryant\", \"isbn\" : \"9787111321330\", \"publishedDate\" : \"2011-01-01\", \"price\" : 99.0, \"stock\" : 50 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}