# book-server/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  contact:
    email: team@example.com
    name: 开发团队
  description: 图书管理系统 API 规范
  title: Bookstore API
  version: 1.0.0
servers:
- description: 本地开发服务器
  url: http://localhost:8080/api
paths:
  /books:
    get:
      description: 返回系统中的所有图书列表
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Book'
                type: array
          description: 成功获取图书列表
        "500":
          description: 服务器内部错误
      summary: 获取所有图书
      x-accepts:
      - application/json
    post:
      description: 向系统中添加一本新图书
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookInput'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
          description: 图书创建成功
        "400":
          description: 请求参数无效
        "500":
          description: 服务器内部错误
      summary: 添加新图书
      x-content-type: application/json
      x-accepts:
      - application/json
  /books/{bookId}:
    delete:
      description: 删除指定ID的图书
      parameters:
      - explode: false
        in: path
        name: bookId
        required: true
        schema:
          minimum: 1
          type: integer
        style: simple
      responses:
        "204":
          description: 图书删除成功
        "404":
          description: 图书不存在
        "500":
          description: 服务器内部错误
      summary: 删除图书
      x-accepts:
      - application/json
    get:
      description: 返回指定ID的图书详情
      parameters:
      - description: 图书ID
        explode: false
        in: path
        name: bookId
        required: true
        schema:
          minimum: 1
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
          description: 成功获取图书
        "404":
          description: 图书不存在
        "500":
          description: 服务器内部错误
      summary: 根据ID获取图书
      x-accepts:
      - application/json
    put:
      description: 更新指定ID的图书信息
      parameters:
      - explode: false
        in: path
        name: bookId
        required: true
        schema:
          minimum: 1
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookInput'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
          description: 图书更新成功
        "404":
          description: 图书不存在
        "400":
          description: 请求参数无效
        "500":
          description: 服务器内部错误
      summary: 更新图书信息
      x-content-type: application/json
      x-accepts:
      - application/json
components:
  schemas:
    Book:
      example:
        id: 1
        title: 深入理解计算机系统
        author: Randal E. Bryant
        isbn: "9787111321330"
        publishedDate: 2011-01-01
        price: 99.0
        stock: 50
      properties:
        id:
          description: 图书唯一标识符
          format: int64
          minimum: 1
          type: integer
        title:
          description: 图书标题
          maxLength: 200
          minLength: 1
          type: string
        author:
          description: 作者姓名
          maxLength: 100
          minLength: 1
          type: string
        isbn:
          description: ISBN编号（13位数字）
          pattern: "^[0-9]{13}$"
          type: string
        publishedDate:
          description: 出版日期
          format: date
          type: string
        price:
          description: 图书价格
          format: float
          minimum: 0
          type: number
        stock:
          description: 库存数量
          minimum: 0
          type: integer
      required:
      - author
      - id
      - title
      type: object
    BookInput:
      example:
        title: 代码大全
        author: Steve McConnell
        isbn: "9787121022982"
        publishedDate: 2006-03-01
        price: 88.0
        stock: 100
      properties:
        title:
          description: 图书标题
          maxLength: 200
          minLength: 1
          type: string
        author:
          description: 作者姓名
          maxLength: 100
          minLength: 1
          type: string
        isbn:
          description: ISBN编号（13位数字）
          pattern: "^[0-9]{13}$"
          type: string
        publishedDate:
          description: 出版日期
          format: date
          type: string
        price:
          description: 图书价格
          format: float
          minimum: 0
          type: number
        stock:
          description: 库存数量
          minimum: 0
          type: integer
      required:
      - author
      - title
      type: object